feat: refactor command to run as std::variant instead

+++ storage/lru_cache.hpp
@@ -51,7 +51,7 @@ class LRUCache {
     ~LRUCache();
     LRUCache(const LRUCache&) = delete;
     LRUCache& operator=(const LRUCache&) = delete;
-    void add(const K& key, const V& value, uint8_t ttl_ms = 0);
+    void add(const K& key, const V& value, uint32_t ttl_ms = 0);
     V& get(const K& key);
     void remove(const K& key);
     void remove_expired();
